get rid of cast warnings

+++ src/tls_dummy.c
@@ -3,7 +3,7 @@
 #if !MG_ENABLE_MBEDTLS && !MG_ENABLE_OPENSSL && !MG_ENABLE_CUSTOM_TLS
 void* mg_tls_ctx_init(const struct mg_tls_opts *opts) {
   (void) opts;
-  return (void *)mg_tls_ctx_init;
+  return (void *)(size_t)mg_tls_ctx_init;
 }
 void mg_tls_ctx_free(void *ctx) {
   (void) ctx;
